# services/api-service/src/generated/java/de/tum/devops/vibeshield/generated/api/ScansApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.tum.devops.vibeshield.generated.api;

import de.tum.devops.vibeshield.generated.model.Error;
import de.tum.devops.vibeshield.generated.model.Finding;
import de.tum.devops.vibeshield.generated.model.Scan;
import de.tum.devops.vibeshield.generated.model.ScanComparison;
import de.tum.devops.vibeshield.generated.model.ScanRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "Scans", description = "Scan jobs and their findings (api-service).")
public interface ScansApi {

    /**
     * GET /api/v1/websites/{websiteId}/scans/latest : Return the most recent scan of a website
     *
     * @param websiteId ID of a website registered by the authenticated user. (required)
     * @return The most recently created scan. (status code 200)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Website not visible to this user, or it has no scans yet (&#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     */
    @Operation(
        operationId = "getLatestScan",
        summary = "Return the most recent scan of a website",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The most recently created scan.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Scan.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Website not visible to this user, or it has no scans yet (`SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/websites/{websiteId}/scans/latest",
        produces = { "application/json" }
    )
    
    ResponseEntity<Scan> getLatestScan(
        @Parameter(name = "websiteId", description = "ID of a website registered by the authenticated user.", required = true, in = ParameterIn.PATH) @PathVariable("websiteId") Long websiteId
    );


    /**
     * GET /api/v1/scans/{scanId} : Return a scan&#39;s current status
     * Poll this endpoint while the status is &#x60;Pending&#x60; or &#x60;Running&#x60;.
     *
     * @param scanId ID of a scan belonging to one of the user&#39;s websites. (required)
     * @return Current scan state. (status code 200)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     */
    @Operation(
        operationId = "getScan",
        summary = "Return a scan's current status",
        description = "Poll this endpoint while the status is `Pending` or `Running`.",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Current scan state.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Scan.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/scans/{scanId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Scan> getScan(
        @Parameter(name = "scanId", description = "ID of a scan belonging to one of the user's websites.", required = true, in = ParameterIn.PATH) @PathVariable("scanId") Long scanId
    );


    /**
     * GET /api/v1/scans/{scanId}/comparison : Compare this scan with the previous completed scan
     * Compares the selected completed scan with the latest previous completed scan for the same website.
     *
     * @param scanId ID of a scan belonging to one of the user&#39;s websites. (required)
     * @return Comparison result and prioritized action plan. (status code 200)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     */
    @Operation(
        operationId = "getScanComparison",
        summary = "Compare this scan with the previous completed scan",
        description = "Compares the selected completed scan with the latest previous completed scan for the same website.",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Comparison result and prioritized action plan.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ScanComparison.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/scans/{scanId}/comparison",
        produces = { "application/json" }
    )
    
    ResponseEntity<ScanComparison> getScanComparison(
        @Parameter(name = "scanId", description = "ID of a scan belonging to one of the user's websites.", required = true, in = ParameterIn.PATH) @PathVariable("scanId") Long scanId
    );


    /**
     * GET /api/v1/scans/{scanId}/findings : List the findings of a scan
     * Empty until the scan status is &#x60;Completed&#x60;.
     *
     * @param scanId ID of a scan belonging to one of the user&#39;s websites. (required)
     * @return Findings ordered by severity, most severe first. (status code 200)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     */
    @Operation(
        operationId = "listFindings",
        summary = "List the findings of a scan",
        description = "Empty until the scan status is `Completed`.",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Findings ordered by severity, most severe first.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Finding.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/scans/{scanId}/findings",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<Finding>> listFindings(
        @Parameter(name = "scanId", description = "ID of a scan belonging to one of the user's websites.", required = true, in = ParameterIn.PATH) @PathVariable("scanId") Long scanId
    );


    /**
     * GET /api/v1/websites/{websiteId}/scans : List the scan history of a website
     *
     * @param websiteId ID of a website registered by the authenticated user. (required)
     * @return All scans for the website, newest first. (status code 200)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     */
    @Operation(
        operationId = "listScans",
        summary = "List the scan history of a website",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "All scans for the website, newest first.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Scan.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/websites/{websiteId}/scans",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<Scan>> listScans(
        @Parameter(name = "websiteId", description = "ID of a website registered by the authenticated user.", required = true, in = ParameterIn.PATH) @PathVariable("websiteId") Long websiteId
    );


    /**
     * POST /api/v1/scans/{scanId}/rescan : Rerun a scan using the same configuration
     * Creates a new Pending scan for the same website and scan configuration as the selected scan.
     *
     * @param scanId ID of a scan belonging to one of the user&#39;s websites. (required)
     * @return Rescan accepted for background execution. (status code 202)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     *         or A scan for this website is already &#x60;Pending&#x60; or &#x60;Running&#x60; (&#x60;SCAN_IN_PROGRESS&#x60;). (status code 409)
     */
    @Operation(
        operationId = "rescan",
        summary = "Rerun a scan using the same configuration",
        description = "Creates a new Pending scan for the same website and scan configuration as the selected scan.",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Rescan accepted for background execution.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Scan.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "409", description = "A scan for this website is already `Pending` or `Running` (`SCAN_IN_PROGRESS`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/scans/{scanId}/rescan",
        produces = { "application/json" }
    )
    
    ResponseEntity<Scan> rescan(
        @Parameter(name = "scanId", description = "ID of a scan belonging to one of the user's websites.", required = true, in = ParameterIn.PATH) @PathVariable("scanId") Long scanId
    );


    /**
     * POST /api/v1/websites/{websiteId}/scans : Trigger a new scan for a website
     * Creates the scan in status &#x60;Pending&#x60; and returns immediately; execution happens in the background. At most one scan per website may be &#x60;Pending&#x60; or &#x60;Running&#x60; at a time — a second trigger returns &#x60;409&#x60;. 
     *
     * @param websiteId ID of a website registered by the authenticated user. (required)
     * @param scanRequest Optional scan configuration; omitting it runs all checks with defaults. (optional)
     * @return Scan accepted for background execution. (status code 202)
     *         or Missing, expired, or invalid Bearer token (&#x60;UNAUTHORIZED&#x60;, &#x60;INVALID_TOKEN&#x60;). (status code 401)
     *         or Resource does not exist or belongs to another user (&#x60;WEBSITE_NOT_FOUND&#x60;, &#x60;SCAN_NOT_FOUND&#x60;). (status code 404)
     *         or A scan for this website is already &#x60;Pending&#x60; or &#x60;Running&#x60; (&#x60;SCAN_IN_PROGRESS&#x60;). (status code 409)
     */
    @Operation(
        operationId = "triggerScan",
        summary = "Trigger a new scan for a website",
        description = "Creates the scan in status `Pending` and returns immediately; execution happens in the background. At most one scan per website may be `Pending` or `Running` at a time — a second trigger returns `409`. ",
        tags = { "Scans" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Scan accepted for background execution.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Scan.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "409", description = "A scan for this website is already `Pending` or `Running` (`SCAN_IN_PROGRESS`).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/websites/{websiteId}/scans",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Scan> triggerScan(
        @Parameter(name = "websiteId", description = "ID of a website registered by the authenticated user.", required = true, in = ParameterIn.PATH) @PathVariable("websiteId") Long websiteId,
        @Parameter(name = "ScanRequest", description = "Optional scan configuration; omitting it runs all checks with defaults.") @Valid @RequestBody(required = false) @Nullable ScanRequest scanRequest
    );

}